- now we are able to express kind of stack alloc impl and number of reserved pages through this_fiber_t - by means of previous point, now co iterators use a specializable struct to resolve kind of stack alloc impl and number of pages to reserve - fixed release compilation errors (pending to investigate why release version does not work properly with fibers) - added attach version to ddk_sync_executor so clients can pass directly a sync executor impl of their own

+++ src/ddkFramework/ddk_co_iterator_context.cpp
@@ -92,7 +92,7 @@ size_t co_random_access_iterator_context::get_curr_index() const
 }
 int co_random_access_iterator_context::shift() const
 {
-	return m_impl.get_target() - m_impl.get_current();
+	return static_cast<int>(m_impl.get_target()) - static_cast<int>(m_impl.get_current());
 }
 
 }
